+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useCamera';
 export * from './useAudioCoach';
 export * from './useMedicalProfile';
 export * from './usePoseTracking';
+export * from './useExerciseTracking';